refactor(import): extract XmlLoader from AegisXmlParser
Move file loading — XML_MAX_BYTES size cap, BOM stripping (UTF-8/UTF-16),
DOCTYPE/XXE rejection, and the LIBXML_NOCDATA|LIBXML_NONET parse — into
NwsCad\Import\XmlLoader (constructed with the PSR-3 logger). loadXmlFile()
becomes a thin delegating wrapper and stripBOM() moves into the loader.
Repurpose XmlLoadGuardsTest to cover XmlLoader directly (adds a BOM case).

Also backfill @uses \NwsCad\Db\UpsertBuilder / \NwsCad\Api\DbHelper on the
call-ingesting tests — the prior UpsertBuilder wiring only surfaces those as
strict-coverage risky under the pcov driver (CI's coverage run would have
flagged them; local --no-coverage runs did not).

Behavior-preserving: coverage set green under pcov (42 tests, no risky) on
MySQL; characterization + load guards green on PostgreSQL (11 tests/132).

Part of #49.
