+++ tests/Unit/XmlLoadGuardsTest.php
@@ -4,24 +4,18 @@ declare(strict_types=1);
 
 namespace NwsCad\Tests\Unit;
 
-use NwsCad\AegisXmlParser;
+use NwsCad\Import\XmlLoader;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
-use ReflectionClass;
-use ReflectionMethod;
 use SimpleXMLElement;
 
 /**
- * Characterizes AegisXmlParser::loadXmlFile()'s pre-parse guards — the
- * XML_MAX_BYTES size cap and the DOCTYPE (XXE) rejection — without a database.
- * loadXmlFile() only touches the filesystem, libxml, and the logger, so we run
- * it on a constructor-less instance with a NullLogger injected via reflection.
+ * Tests XmlLoader's pre-parse guards — the XML_MAX_BYTES size cap, BOM
+ * stripping, and the DOCTYPE/XXE rejection (#49, extracted from
+ * AegisXmlParser::loadXmlFile). No database; a NullLogger stands in for the
+ * PSR-3 logger and the oversized input is generated at runtime.
  *
- * Covers the "oversized file" and XXE fixture cases from #48 without committing
- * a multi-megabyte artifact to the repo (the oversized input is generated at
- * runtime against a lowered XML_MAX_BYTES cap).
- *
- * @covers \NwsCad\AegisXmlParser
+ * @covers \NwsCad\Import\XmlLoader
  */
 class XmlLoadGuardsTest extends TestCase
 {
@@ -56,19 +50,9 @@ class XmlLoadGuardsTest extends TestCase
         return $path;
     }
 
-    private function loadXml(string $path): SimpleXMLElement|false
+    private function load(string $path): SimpleXMLElement|false
     {
-        $rc = new ReflectionClass(AegisXmlParser::class);
-        $obj = $rc->newInstanceWithoutConstructor();
-        $logger = $rc->getProperty('logger');
-        $logger->setAccessible(true);
-        $logger->setValue($obj, new NullLogger());
-
-        $m = new ReflectionMethod(AegisXmlParser::class, 'loadXmlFile');
-        $m->setAccessible(true);
-        /** @var SimpleXMLElement|false $result */
-        $result = $m->invoke($obj, $path);
-        return $result;
+        return (new XmlLoader(new NullLogger()))->load($path);
     }
 
     public function testOversizedFileIsRejected(): void
@@ -80,7 +64,7 @@ class XmlLoadGuardsTest extends TestCase
         $path = $this->writeTmp($body);
         $this->assertGreaterThan(1024, strlen($body), 'precondition: input exceeds the cap');
 
-        $this->assertFalse($this->loadXml($path), 'file over XML_MAX_BYTES must be rejected');
+        $this->assertFalse($this->load($path), 'file over XML_MAX_BYTES must be rejected');
     }
 
     public function testFileWithinCapLoads(): void
@@ -91,7 +75,7 @@ class XmlLoadGuardsTest extends TestCase
             . '<CallExport xmlns="http://www.newworldsystems.com/Aegis/CAD/Peripheral/CallExport/2011/02">'
             . '<CallId>1</CallId></CallExport>'
         );
-        $this->assertInstanceOf(SimpleXMLElement::class, $this->loadXml($path));
+        $this->assertInstanceOf(SimpleXMLElement::class, $this->load($path));
     }
 
     public function testDoctypeDeclarationIsRejected(): void
@@ -101,6 +85,24 @@ class XmlLoadGuardsTest extends TestCase
             . "<!DOCTYPE foo [<!ENTITY xxe SYSTEM \"file:///etc/passwd\">]>\n"
             . '<CallExport><CallId>&xxe;</CallId></CallExport>'
         );
-        $this->assertFalse($this->loadXml($path), 'a DOCTYPE declaration must be rejected outright');
+        $this->assertFalse($this->load($path), 'a DOCTYPE declaration must be rejected outright');
+    }
+
+    public function testMissingFileIsRejected(): void
+    {
+        $this->assertFalse($this->load(sys_get_temp_dir() . '/does_not_exist_' . uniqid() . '.xml'));
+    }
+
+    public function testUtf8BomIsStrippedAndParsed(): void
+    {
+        $path = $this->writeTmp(
+            "\xEF\xBB\xBF"
+            . '<?xml version="1.0" encoding="UTF-8"?>'
+            . '<CallExport xmlns="http://www.newworldsystems.com/Aegis/CAD/Peripheral/CallExport/2011/02">'
+            . '<CallId>42</CallId></CallExport>'
+        );
+        $xml = $this->load($path);
+        $this->assertInstanceOf(SimpleXMLElement::class, $xml);
+        $this->assertSame('42', (string) $xml->CallId);
     }
 }
